fix: returning the same metric samples multiple times
The old '_getMetricValues' appended metrics via 'self.metrics["<name>"].add_metric(…)'
but then 'returned' all metrics via 'yield self.metrics["<name>"]' for every device.

So when calling '_getMetricValues' for the first device it returns only the metrics for
the first device, but when calling it again (for the second device) it returns the
metrics for the first **and** second device (and so forth).

To fix this 'FritzCapability._getMetricValues()' is split into '_generateMetricValues()'
and '_getMetricValues()'.
With most of the code staying in '_generateMetricValues()' except for the 'yield'
statements which are moved into the new '_getMetricValues()'.

With this 'FritzCapability.getMetrics()' can stay largely the same with the generating
of the metric values staying in the 'for … devices:' loop, and only the actual retrieving
of the metric values moved out of the loop.